requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alembic==1.16.2
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
attrs==25.3.0
audioread==3.0.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorlog==6.9.0
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
docopt==0.6.2
einops==0.8.1
filelock==3.18.0
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
hf-xet==1.1.5
huggingface-hub==0.33.2
HyperPyYAML==1.2.2
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
julius==0.2.7
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
lightning==2.5.2
lightning-utilities==0.14.3
llvmlite==0.44.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
networkx==3.5
numba==0.61.2
numpy==2.2.6
omegaconf==2.3.0
optuna==4.4.0
packaging==25.0
pandas==2.3.0
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
primePy==1.3
propcache==0.3.2
protobuf==6.31.1
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.22
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytorch-lightning==2.5.2
pytorch-metric-learning==2.8.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
scikit-learn==1.7.0
scipy==1.16.0
semver==3.0.4
sentencepiece==0.2.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==0.5.0.post1
speechbrain==1.0.3
SQLAlchemy==2.0.41
sympy==1.14.0
tabulate==0.9.0
tensorboardX==2.6.4
threadpoolctl==3.6.0
torch==2.7.1
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.7.1
torchmetrics==1.7.3
tqdm==4.67.1
typer==0.16.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
yarl==1.20.1
